+++ src/pudl/transform/params/ferc1.py
@@ -1,0 +1,1786 @@
+"""FERC 1 specific transformation parameters."""
+
+from datetime import date
+
+##############################################################################
+# Unit converstion parameters
+##############################################################################
+PERPOUND_TO_PERSHORTTON = dict(
+    multiplier=2000.0,
+    pattern=r"(.*)_per_lb$",
+    repl=r"\1_per_ton",
+)
+"""Parameters for converting from inverse pounds to inverse short tons."""
+
+CENTS_TO_DOLLARS = dict(
+    multiplier=0.01,
+    pattern=r"(.*)_cents$",
+    repl=r"\1_usd",
+)
+"""Parameters for converting from cents to dollars."""
+
+PERCF_TO_PERMCF = dict(
+    multiplier=1000.0,
+    pattern=r"(.*)_per_cf$",
+    repl=r"\1_per_mcf",
+)
+"""Parameters for converting from inverse cubic feet to inverse 1000s of cubic feet."""
+
+PERGALLON_TO_PERBARREL = dict(
+    multiplier=42.0,
+    pattern=r"(.*)_per_gal",
+    repl=r"\1_per_bbl",
+)
+"""Parameters for converting from inverse gallons to inverse barrels."""
+
+PERKW_TO_PERMW = dict(
+    multiplier=1000.0,
+    pattern=r"(.*)_per_kw$",
+    repl=r"\1_per_mw",
+)
+"""Parameters for converting column units from per kW to per MW."""
+
+PERKWH_TO_PERMWH = dict(
+    multiplier=1000.0,
+    pattern=r"(.*)_per_kwh$",
+    repl=r"\1_per_mwh",
+)
+"""Parameters for converting column units from per kWh to per MWh."""
+
+KWH_TO_MWH = dict(
+    multiplier=1e-3,
+    pattern=r"(.*)_kwh$",
+    repl=r"\1_mwh",
+)
+"""Parameters for converting column units from kWh to MWh."""
+
+BTU_TO_MMBTU = dict(
+    multiplier=1e-6,
+    pattern=r"(.*)_btu(.*)$",
+    repl=r"\1_mmbtu\2",
+)
+"""Parameters for converting column units from BTU to MMBTU."""
+
+PERBTU_TO_PERMMBTU = dict(
+    multiplier=1e6,
+    pattern=r"(.*)_per_btu$",
+    repl=r"\1_per_mmbtu",
+)
+"""Parameters for converting column units from BTU to MMBTU."""
+
+BTU_PERKWH_TO_MMBTU_PERMWH = dict(
+    multiplier=(1e-6 * 1000.0),
+    pattern=r"(.*)_btu_per_kwh$",
+    repl=r"\1_mmbtu_per_mwh",
+)
+"""Parameters for converting column units from BTU/kWh to MMBTU/MWh."""
+
+
+##############################################################################
+# Valid ranges to impose on various columns
+##############################################################################
+VALID_PLANT_YEARS = {
+    "lower_bound": 1850,
+    "upper_bound": date.today().year,
+}
+"""Valid range of years for power plant construction."""
+
+VALID_COAL_MMBTU_PER_TON = {
+    "lower_bound": 6.5,
+    "upper_bound": 29.0,
+}
+"""Valid range for coal heat content, taken from the EIA-923 instructions.
+
+Lower bound is for waste coal. Upper bound is for bituminous coal.
+
+https://www.eia.gov/survey/form/eia_923/instructions.pdf
+"""
+
+VALID_COAL_USD_PER_MMBTU = {
+    "lower_bound": 0.5,
+    "upper_bound": 7.5,
+}
+"""Historical coal price range from the EIA-923 Fuel Receipts and Costs table."""
+
+VALID_GAS_MMBTU_PER_MCF = {
+    "lower_bound": 0.3,
+    "upper_bound": 3.3,
+}
+"""Valid range for gaseous fuel heat content, taken from the EIA-923 instructions.
+
+Lower bound is for landfill gas. Upper bound is for "other gas".  Blast furnace gas
+(which has very low heat content) is effectively excluded.
+
+https://www.eia.gov/survey/form/eia_923/instructions.pdf
+"""
+
+VALID_GAS_USD_PER_MMBTU = {
+    "lower_bound": 1.0,
+    "upper_bound": 35.0,
+}
+"""Historical natural gas price range from the EIA-923 Fuel Receipts and Costs table."""
+
+VALID_OIL_MMBTU_PER_BBL = {
+    "lower_bound": 3.0,
+    "upper_bound": 6.9,
+}
+"""Valid range for petroleum fuels heat content, taken from the EIA-923 instructions.
+
+Lower bound is for waste oil. Upper bound is for residual fuel oil.
+
+https://www.eia.gov/survey/form/eia_923/instructions.pdf
+"""
+
+VALID_OIL_USD_PER_MMBTU = {
+    "lower_bound": 5.0,
+    "upper_bound": 33.0,
+}
+"""Historical petroleum price range from the EIA-923 Fuel Receipts and Costs table."""
+
+##############################################################################
+# String categorizations
+##############################################################################
+FUEL_CATEGORIES: dict[str, set[str]] = {
+    "categories": {
+        "coal": {
+            "coal",
+            "coal-subbit",
+            "lignite",
+            "coal(sb)",
+            "coal (sb)",
+            "coal-lignite",
+            "coke",
+            "coa",
+            "lignite/coal",
+            "coal - subbit",
+            "coal-subb",
+            "coal-sub",
+            "coal-lig",
+            "coal-sub bit",
+            "coals",
+            "ciak",
+            "petcoke",
+            "coal.oil",
+            "coal/gas",
+            "coal. gas",
+            "coal & oil",
+            "coal bit",
+            "bit coal",
+            "coal-unit #3",
+            "coal-subbitum",
+            "coal tons",
+            "coal mcf",
+            "coal unit #3",
+            "pet. coke",
+            "coal-u3",
+            "coal&coke",
+            "tons",
+        },
+        "oil": {
+            "oil",
+            "#6 oil",
+            "#2 oil",
+            "fuel oil",
+            "jet",
+            "no. 2 oil",
+            "no.2 oil",
+            "no.6& used",
+            "used oil",
+            "oil-2",
+            "oil (#2)",
+            "diesel oil",
+            "residual oil",
+            "# 2 oil",
+            "resid. oil",
+            "tall oil",
+            "oil/gas",
+            "no.6 oil",
+            "oil-fuel",
+            "oil-diesel",
+            "oil / gas",
+            "oil bbls",
+            "oil bls",
+            "no. 6 oil",
+            "#1 kerosene",
+            "diesel",
+            "no. 2 oils",
+            "blend oil",
+            "#2oil diesel",
+            "#2 oil-diesel",
+            "# 2  oil",
+            "light oil",
+            "heavy oil",
+            "gas.oil",
+            "#2",
+            "2",
+            "6",
+            "bbl",
+            "no 2 oil",
+            "no 6 oil",
+            "#1 oil",
+            "#6",
+            "oil-kero",
+            "oil bbl",
+            "biofuel",
+            "no 2",
+            "kero",
+            "#1 fuel oil",
+            "no. 2  oil",
+            "blended oil",
+            "no 2. oil",
+            "# 6 oil",
+            "nno. 2 oil",
+            "#2 fuel",
+            "oill",
+            "oils",
+            "gas/oil",
+            "no.2 oil gas",
+            "#2 fuel oil",
+            "oli",
+            "oil (#6)",
+            "oil/diesel",
+            "2 oil",
+            "#6 hvy oil",
+            "jet fuel",
+            "diesel/compos",
+            "oil-8",
+            "oil {6}",  # noqa: FS003
+            "oil-unit #1",
+            "bbl.",
+            "oil.",
+            "oil #6",
+            "oil (6)",
+            "oil(#2)",
+            "oil-unit1&2",
+            "oil-6",
+            "#2 fue oil",
+            "dielel oil",
+            "dielsel oil",
+            "#6 & used",
+            "barrels",
+            "oil un 1 & 2",
+            "jet oil",
+            "oil-u1&2",
+            "oiul",
+            "pil",
+            "oil - 2",
+            "#6 & used",
+            "oial",
+            "diesel fuel",
+            "diesel/compo",
+            "oil (used)",
+        },
+        "gas": {
+            "gas",
+            "gass",
+            "methane",
+            "natural gas",
+            "blast gas",
+            "gas mcf",
+            "propane",
+            "prop",
+            "natural  gas",
+            "nat.gas",
+            "nat gas",
+            "nat. gas",
+            "natl gas",
+            "ga",
+            "gas`",
+            "syngas",
+            "ng",
+            "mcf",
+            "blast gaa",
+            "nat  gas",
+            "gac",
+            "syngass",
+            "prop.",
+            "natural",
+            "coal.gas",
+            "n. gas",
+            "lp gas",
+            "natuaral gas",
+            "coke gas",
+            "gas #2016",
+            "propane**",
+            "* propane",
+            "propane **",
+            "gas expander",
+            "gas ct",
+            "# 6 gas",
+            "#6 gas",
+            "coke oven gas",
+            "gas & oil",
+            "gas/fuel oil",
+        },
+        "solar": {"solar"},
+        "wind": {"wind"},
+        "hydro": {"hydro"},
+        "nuclear": {
+            "nuclear",
+            "grams of uran",
+            "grams of",
+            "grams of  ura",
+            "grams",
+            "nucleur",
+            "nulear",
+            "nucl",
+            "nucleart",
+            "nucelar",
+            "gr.uranium",
+            "grams of urm",
+            "nuclear (9)",
+            "nulcear",
+            "nuc",
+            "gr. uranium",
+            "uranium",
+            "nuclear mw da",
+            "grams of ura",
+            "nucvlear",
+            "nuclear (1)",
+        },
+        "waste": {
+            "waste",
+            "tires",
+            "tire",
+            "refuse",
+            "switchgrass",
+            "wood waste",
+            "woodchips",
+            "biomass",
+            "wood",
+            "wood chips",
+            "rdf",
+            "tires/refuse",
+            "tire refuse",
+            "waste oil",
+            "woodships",
+            "tire chips",
+            "tdf",
+        },
+        "other": {  # This should really be NA but we require a fuel_type_code_pudl
+            "other",
+            "na",
+            "",
+            "steam",
+            "purch steam",
+            "all",
+            "n/a",
+            "purch. steam",
+            "composite",
+            "composit",
+            "mbtus",
+            "total",
+            "avg",
+            "avg.",
+            "blo",
+            "all fuel",
+            "comb.",
+            "alt. fuels",
+            "comb",
+            "/#=2\x80â\x91?",
+            "kã\xadgv¸\x9d?",
+            "mbtu's",
+            "gas, oil",
+            "rrm",
+            "3\x9c",
+            "average",
+            "furfural",
+            "0",
+            "watson bng",
+            "toal",
+            "bng",
+            "# 6 & used",
+            "combined",
+            "blo bls",
+            "compsite",
+            "*",
+            "compos.",
+            "gas / oil",
+            "mw days",
+            "g",
+            "c",
+            "lime",
+            "all fuels",
+            "at right",
+            "20",
+            "1",
+            "comp oil/gas",
+            "all fuels to",
+            "the right are",
+            "c omposite",
+            "all fuels are",
+            "total pr crk",
+            "all fuels =",
+            "total pc",
+            "comp",
+            "alternative",
+            "alt. fuel",
+            "bio fuel",
+            "total prairie",
+            "kã\xadgv¸?",
+            "m",
+            "waste heat",
+            "/#=2â?",
+            "3",
+            "—",
+        },
+    }
+}
+"""
+A mapping a canonical fuel name to a set of strings which are used to represent that
+fuel in the FERC Form 1 Reporting. Case is ignored, as all fuel strings are converted to
+lower case in the data set.
+"""
+
+FUEL_UNIT_CATEGORIES: dict[str, set[str]] = {
+    "categories": {
+        "ton": {
+            "ton",
+            "toms",
+            "taons",
+            "tones",
+            "col-tons",
+            "toncoaleq",
+            "coal",
+            "tons coal eq",
+            "coal-tons",
+            "tons",
+            "tons coal",
+            "coal-ton",
+            "tires-tons",
+            "coal tons -2 ",
+            "oil-tons",
+            "coal tons 200",
+            "ton-2000",
+            "coal tons",
+            "coal tons -2",
+            "coal-tone",
+            "tire-ton",
+            "tire-tons",
+            "ton coal eqv",
+            "tos",
+            "coal tons - 2",
+            "c. t.",
+            "c.t.",
+            "t",
+            "toncoalequiv",
+        },
+        "mcf": {
+            "mcf",
+            "mcf's",
+            "mcfs",
+            "mcf.",
+            "mcfe",
+            "gas mcf",
+            '"gas" mcf',
+            "gas-mcf",
+            "mfc",
+            "mct",
+            " mcf",
+            "msfs",
+            "mlf",
+            "mscf",
+            "mci",
+            "mcl",
+            "mcg",
+            "m.cu.ft.",
+            "kcf",
+            "(mcf)",
+            "mcf *(4)",
+            "mcf00",
+            "m.cu.ft..",
+            "1000 c.f",
+        },
+        "bbl": {
+            "bbl",
+            "barrel",
+            "bbls",
+            "barrels",
+            "bbrl",
+            "bbl.",
+            "bbls.",
+            "oil 42 gal",
+            "oil-barrels",
+            "barrrels",
+            "bbl-42 gal",
+            "oil-barrel",
+            "bb.",
+            "barrells",
+            "bar",
+            "bbld",
+            "oil- barrel",
+            "barrels    .",
+            "bbl .",
+            "barels",
+            "barrell",
+            "berrels",
+            "bb",
+            "bbl.s",
+            "oil-bbl",
+            "bls",
+            "bbl:",
+            "barrles",
+            "blb",
+            "propane-bbl",
+            "barriel",
+            "berriel",
+            "barrile",
+            "(bbl.)",
+            "barrel *(4)",
+            "(4) barrel",
+            "bbf",
+            "blb.",
+            "(bbl)",
+            "bb1",
+            "bbsl",
+            "barrrel",
+            "barrels 100%",
+            "bsrrels",
+            "bbl's",
+            "*barrels",
+            "oil - barrels",
+            "oil 42 gal ba",
+            "bll",
+            "boiler barrel",
+            "gas barrel",
+            '"boiler" barr',
+            '"gas" barrel',
+            '"boiler"barre',
+            '"boiler barre',
+            "barrels .",
+            "bariel",
+            "brrels",
+            "oil barrel",
+            "barreks",
+            "oil-bbls",
+            "oil-bbs",
+            "boe",
+        },
+        "mmbbl": {"mmbbl", "mmbbls"},
+        "gal": {"gal", "gallons", "gal.", "gals", "gals.", "gallon", "galllons"},
+        "kgal": {
+            "kgal",
+            "oil(1000 gal)",
+            "oil(1000)",
+            "oil (1000)",
+            "oil(1000",
+            "oil(1000ga)",
+            "1000 gals",
+            "1000 gal",
+        },
+        "grams": {
+            "gram",
+            "grams",
+            "gm u",
+            "grams u235",
+            "grams u-235",
+            "grams of uran",
+            "grams: u-235",
+            "grams:u-235",
+            "grams:u235",
+            "grams u308",
+            "grams: u235",
+            "grams of",
+            "grams - n/a",
+            "gms uran",
+            "s e uo2 grams",
+            "gms uranium",
+            "grams of urm",
+            "gms. of uran",
+            "grams (100%)",
+            "grams v-235",
+            "se uo2 grams",
+            "grams u",
+            "g",
+            "grams of uranium",
+        },
+        "kg": {
+            "kg",
+            "kg of uranium",
+            "kg uranium",
+            "kilg. u-235",
+            "kg u-235",
+            "kilograms-u23",
+            "kilograms u-2",
+            "kilograms",
+            "kg of",
+            "kg-u-235",
+            "kilgrams",
+            "kilogr. u235",
+            "uranium kg",
+            "kg uranium25",
+            "kilogr. u-235",
+            "kg uranium 25",
+            "kilgr. u-235",
+            "kguranium 25",
+            "kg-u235",
+            "kgm",
+        },
+        "klbs": {
+            "klbs",
+            "k lbs.",
+            "k lbs",
+            "1000 / lbs",
+            "1000 lbs",
+        },
+        "mmbtu": {
+            "mmbtu",
+            "mmbtus",
+            "mbtus",
+            "(mmbtu)",
+            "mmbtu's",
+            "nuclear-mmbtu",
+            "nuclear-mmbt",
+            "mmbtul",
+        },
+        "btu": {
+            "btu",
+            "btus",
+        },
+        "mwdth": {
+            "mwdth",
+            "mwd therman",
+            "mw days-therm",
+            "mwd thrml",
+            "mwd thermal",
+            "mwd/mtu",
+            "mw days",
+            "mwd",
+            "mw day",
+            "dth",
+            "mwdaysthermal",
+            "mw day therml",
+            "mw days thrml",
+            "nuclear mwd",
+            "mmwd",
+            "mw day/therml" "mw days/therm",
+            "mw days (th",
+            "ermal)",
+        },
+        "mwhth": {
+            "nwh therm",
+            "mwhth",
+            "mwh them",
+            "mwh threm",
+            "mwh therm",
+            "mwh",
+            "mwh therms.",
+            "mwh term.uts",
+            "mwh thermal",
+            "mwh thermals",
+            "mw hr therm",
+            "mwh therma",
+            "mwh therm.uts",
+        },
+        "na_category": {
+            "na_category",
+            "na",
+            "",
+            "1265",
+            "mwh units",
+            "composite",
+            "therms",
+            "n/a",
+            "mbtu/kg",
+            "uranium 235",
+            "oil",
+            "ccf",
+            "2261",
+            "uo2",
+            "(7)",
+            "oil #2",
+            "oil #6",
+            '\x99å\x83\x90?"',
+            "dekatherm",
+            "0",
+            "mw day/therml",
+            "nuclear",
+            "gas",
+            "62,679",
+            "mw days/therm",
+            "uranium",
+            "oil/gas",
+            "thermal",
+            "(thermal)",
+            "se uo2",
+            "181679",
+            "83",
+            "3070",
+            "248",
+            "273976",
+            "747",
+            "-",
+            "are total",
+            "pr. creek",
+            "decatherms",
+            "uramium",
+            ".",
+            "total pr crk",
+            ">>>>>>>>",
+            "all",
+            "total",
+            "alternative-t",
+            "oil-mcf",
+            "3303671",
+            "929",
+            "7182175",
+            "319",
+            "1490442",
+            "10881",
+            "1363663",
+            "7171",
+            "1726497",
+            "4783",
+            "7800",
+            "12559",
+            "2398",
+            "creek fuels",
+            "propane-barre",
+            "509",
+            "barrels/mcf",
+            "propane-bar",
+            "4853325",
+            "4069628",
+            "1431536",
+            "708903",
+            "mcf/oil (1000",
+            "344",
+            'å?"',
+            "mcf / gallen",
+            "none",
+            "—",
+        },
+    }
+}
+"""
+A mapping of canonical fuel units (keys) to sets of strings representing those
+fuel units (values)
+"""
+
+PLANT_TYPE_CATEGORIES: dict[str, set[str]] = {
+    "categories": {
+        "steam": {
+            "coal",
+            "steam",
+            "steam units 1 2 3",
+            "steam units 4 5",
+            "steam fossil",
+            "steam turbine",
+            "steam a",
+            "steam 100",
+            "steam units 1 2 3",
+            "steams",
+            "steam 1",
+            "steam retired 2013",
+            "stream",
+            "steam units 1,2,3",
+            "steam units 4&5",
+            "steam units 4&6",
+            "steam conventional",
+            "unit total-steam",
+            "unit total steam",
+            "*resp. share steam",
+            "resp. share steam",
+            "steam (see note 1,",
+            "steam (see note 3)",
+            "mpc 50%share steam",
+            "40% share steam" "steam (2)",
+            "steam (3)",
+            "steam (4)",
+            "steam (5)",
+            "steam (6)",
+            "steam (7)",
+            "steam (8)",
+            "steam units 1 and 2",
+            "steam units 3 and 4",
+            "steam (note 1)",
+            "steam (retired)",
+            "steam (leased)",
+            "coal-fired steam",
+            "oil-fired steam",
+            "steam/fossil",
+            "steam (a,b)",
+            "steam (a)",
+            "stean",
+            "steam-internal comb",
+            "steam (see notes)",
+            "steam units 4 & 6",
+            "resp share stm note3",
+            "mpc50% share steam",
+            "mpc40%share steam",
+            "steam - 64%",
+            "steam - 100%",
+            "steam (1) & (2)",
+            "resp share st note3",
+            "mpc 50% shares steam",
+            "steam-64%",
+            "steam-100%",
+            "steam (see note 1)",
+            "mpc 50% share steam",
+            "steam units 1, 2, 3",
+            "steam units 4, 5",
+            "steam (2)",
+            "steam (1)",
+            "steam 4, 5",
+            "steam - 72%",
+            "steam (incl i.c.)",
+            "steam- 72%",
+            "steam;retired - 2013",
+            "respondent's sh.-st.",
+            "respondent's sh-st",
+            "40% share steam",
+            "resp share stm note3",
+            "mpc50% share steam",
+            "resp share st note 3",
+            "\x02steam (1)",
+            "coal fired steam tur",
+            "coal fired steam turbine",
+            "steam- 64%",
+        },
+        "combustion_turbine": {
+            "combustion turbine",
+            "gt",
+            "gas turbine",
+            "gas turbine # 1",
+            "gas turbine",
+            "gas turbine (note 1)",
+            "gas turbines",
+            "simple cycle",
+            "combustion turbine",
+            "comb.turb.peak.units",
+            "gas turbine",
+            "combustion turbine",
+            "com turbine peaking",
+            "gas turbine peaking",
+            "comb turb peaking",
+            "combustine turbine",
+            "comb. turine",
+            "conbustion turbine",
+            "combustine turbine",
+            "gas turbine (leased)",
+            "combustion tubine",
+            "gas turb",
+            "gas turbine peaker",
+            "gtg/gas",
+            "simple cycle turbine",
+            "gas-turbine",
+            "gas turbine-simple",
+            "gas turbine - note 1",
+            "gas turbine #1",
+            "simple cycle",
+            "gasturbine",
+            "combustionturbine",
+            "gas turbine (2)",
+            "comb turb peak units",
+            "jet engine",
+            "jet powered turbine",
+            "*gas turbine",
+            "gas turb.(see note5)",
+            "gas turb. (see note",
+            "combutsion turbine",
+            "combustion turbin",
+            "gas turbine-unit 2",
+            "gas - turbine",
+            "comb turbine peaking",
+            "gas expander turbine",
+            "jet turbine",
+            "gas turbin (lease",
+            "gas turbine (leased",
+            "gas turbine/int. cm",
+            "comb.turb-gas oper.",
+            "comb.turb.gas/oil op",
+            "comb.turb.oil oper.",
+            "jet",
+            "comb. turbine (a)",
+            "gas turb.(see notes)",
+            "gas turb(see notes)",
+            "comb. turb-gas oper",
+            "comb.turb.oil oper",
+            "gas turbin (leasd)",
+            "gas turbne/int comb",
+            "gas turbine (note1)",
+            "combution turbin",
+            "* gas turbine",
+            "add to gas turbine",
+            "gas turbine (a)",
+            "gas turbinint comb",
+            "gas turbine (note 3)",
+            "resp share gas note3",
+            "gas trubine",
+            "*gas turbine(note3)",
+            "gas turbine note 3,6",
+            "gas turbine note 4,6",
+            "gas turbine peakload",
+            "combusition turbine",
+            "gas turbine (lease)",
+            "comb. turb-gas oper.",
+            "combution turbine",
+            "combusion turbine",
+            "comb. turb. oil oper",
+            "combustion burbine",
+            "combustion and gas",
+            "comb. turb.",
+            "gas turbine (lease",
+            "gas turbine (leasd)",
+            "gas turbine/int comb",
+            "*gas turbine(note 3)",
+            "gas turbine (see nos",
+            "i.c.e./gas turbine",
+            "gas turbine/intcomb",
+            "cumbustion turbine",
+            "gas turb, int. comb.",
+            "gas turb, diesel",
+            "gas turb, int. comb",
+            "i.c.e/gas turbine",
+            "diesel turbine",
+            "comubstion turbine",
+            "i.c.e. /gas turbine",
+            "i.c.e/ gas turbine",
+            "i.c.e./gas tubine",
+            "gas turbine; retired",
+        },
+        "combined_cycle": {
+            "Combined cycle",
+            "combined cycle",
+            "combined",
+            "gas & steam turbine",
+            "gas turb. & heat rec",
+            "combined cycle",
+            "com. cyc",
+            "com. cycle",
+            "gas turb-combined cy",
+            "combined cycle ctg",
+            "combined cycle - 40%",
+            "com cycle gas turb",
+            "combined cycle oper",
+            "gas turb/comb. cyc",
+            "combine cycle",
+            "cc",
+            "comb. cycle",
+            "gas turb-combined cy",
+            "steam and cc",
+            "steam cc",
+            "gas steam",
+            "ctg steam gas",
+            "steam comb cycle",
+            "gas/steam comb. cycl",
+            "gas/steam",
+            "gas turbine-combined cycle",
+            "steam (comb. cycle)" "gas turbine/steam",
+            "steam & gas turbine",
+            "gas trb & heat rec",
+            "steam & combined ce",
+            "st/gas turb comb cyc",
+            "gas tur & comb cycl",
+            "combined cycle (a,b)",
+            "gas turbine/ steam",
+            "steam/gas turb.",
+            "steam & comb cycle",
+            "gas/steam comb cycle",
+            "comb cycle (a,b)",
+            "igcc",
+            "steam/gas turbine",
+            "gas turbine / steam",
+            "gas tur & comb cyc",
+            "comb cyc (a) (b)",
+            "comb cycle",
+            "comb cyc",
+            "combined turbine",
+            "combine cycle oper",
+            "comb cycle/steam tur",
+            "cc / gas turb",
+            "steam (comb. cycle)",
+            "steam & cc",
+            "gas turbine/steam",
+            "gas turb/cumbus cycl",
+            "gas turb/comb cycle",
+            "gasturb/comb cycle",
+            "gas turb/cumb. cyc",
+            "igcc/gas turbine",
+            "gas / steam",
+            "ctg/steam-gas",
+            "ctg/steam -gas",
+            "gas fired cc turbine",
+            "combinedcycle",
+            "comb cycle gas turb",
+            "combined cycle opern",
+            "comb. cycle gas turb",
+            "ngcc",
+        },
+        "nuclear": {
+            "nuclear",
+            "nuclear (3)",
+            "steam(nuclear)",
+            "nuclear(see note4)" "nuclear steam",
+            "nuclear turbine",
+            "nuclear - steam",
+            "nuclear (a)(b)(c)",
+            "nuclear (b)(c)",
+            "* nuclear",
+            "nuclear (b) (c)",
+            "nuclear (see notes)",
+            "steam (nuclear)",
+            "* nuclear (note 2)",
+            "nuclear (note 2)",
+            "nuclear (see note 2)",
+            "nuclear(see note4)",
+            "nuclear steam",
+            "nuclear(see notes)",
+            "nuclear-steam",
+            "nuclear (see note 3)",
+        },
+        "geothermal": {"steam - geothermal", "steam_geothermal", "geothermal"},
+        "internal_combustion": {
+            "ic",
+            "internal combustion",
+            "internal comb.",
+            "internl combustion",
+            "diesel turbine",
+            "int combust (note 1)",
+            "int. combust (note1)",
+            "int.combustine",
+            "comb. cyc",
+            "internal comb",
+            "diesel",
+            "diesel engine",
+            "internal combustion",
+            "int combust - note 1",
+            "int. combust - note1",
+            "internal comb recip",
+            "internal combustion reciprocating",
+            "reciprocating engine",
+            "comb. turbine",
+            "internal combust.",
+            "int. combustion (1)",
+            "*int combustion (1)",
+            "*internal combust'n",
+            "internal",
+            "internal comb.",
+            "steam internal comb",
+            "combustion",
+            "int. combustion",
+            "int combust (note1)",
+            "int. combustine",
+            "internl combustion",
+            "*int. combustion (1)",
+            "internal conbustion",
+        },
+        "wind": {
+            "wind",
+            "wind energy",
+            "wind turbine",
+            "wind - turbine",
+            "wind generation",
+            "wind turbin",
+        },
+        "photovoltaic": {
+            "solar photovoltaic",
+            "photovoltaic",
+            "solar",
+            "solar project",
+        },
+        "solar_thermal": {"solar thermal"},
+        "na_category": {
+            "na_category",
+            "na",
+            "",
+            "n/a",
+            "see pgs 402.1-402.3",
+            "see pgs 403.1-403.9",
+            "respondent's share",
+            "--",
+            "—",
+            "footnote",
+            "(see note 7)",
+            "other",
+            "not applicable",
+            "peach bottom",
+            "none.",
+            "fuel facilities",
+            "0",
+            "not in service",
+            "none",
+            "common expenses",
+            "expenses common to",
+            "retired in 1981",
+            "retired in 1978",
+            "unit total (note3)",
+            "unit total (note2)",
+            "resp. share (note2)",
+            "resp. share (note8)",
+            "resp. share (note 9)",
+            "resp. share (note11)",
+            "resp. share (note4)",
+            "resp. share (note6)",
+            "conventional",
+            "expenses commom to",
+            "not in service in",
+            "unit total (note 3)",
+            "unit total (note 2)",
+            "resp. share (note 8)",
+            "resp. share (note 3)",
+            "resp. share note 11",
+            "resp. share (note 4)",
+            "resp. share (note 6)",
+            "(see note 5)",
+            "resp. share (note 2)",
+            "package",
+            "(left blank)",
+            "common",
+            "0.0000",
+            "other generation",
+            "resp share (note 11)",
+            "retired",
+            "storage/pipelines",
+            "sold april 16, 1999",
+            "sold may 07, 1999",
+            "plants sold in 1999",
+            "gas",
+            "not applicable.",
+            "resp. share - note 2",
+            "resp. share - note 8",
+            "resp. share - note 9",
+            "resp share - note 11",
+            "resp. share - note 4",
+            "resp. share - note 6",
+            "plant retired- 2013",
+            "retired - 2013",
+            "resp share - note 5",
+            "resp. share - note 7",
+            "non-applicable",
+            "other generation plt",
+            "combined heat/power",
+            "oil",
+            "fuel oil",
+        },
+    }
+}
+"""
+A mapping from canonical plant kinds (keys) to the associated freeform strings (values)
+identified as being associated with that kind of plant in the FERC Form 1 raw data.
+There are many strings that weren't categorized, Solar and Solar Project were not
+classified as these do not indicate if they are solar thermal or photovoltaic. Variants
+on Steam (e.g. "steam 72" and "steam and gas") were classified based on additional
+research of the plants on the Internet.
+"""
+
+CONSTRUCTION_TYPE_CATEGORIES: dict[str, set[str]] = {
+    "categories": {
+        "outdoor": {
+            "outdoor",
+            "outdoor boiler",
+            "full outdoor",
+            "outdoor boiler",
+            "outdoor boilers",
+            "outboilers",
+            "fuel outdoor",
+            "full outdoor",
+            "outdoors",
+            "outdoor",
+            "boiler outdoor& full",
+            "boiler outdoor&full",
+            "outdoor boiler& full",
+            "full -outdoor",
+            "outdoor steam",
+            "outdoor boiler",
+            "ob",
+            "outdoor automatic",
+            "outdoor repower",
+            "full outdoor boiler",
+            "fo",
+            "outdoor boiler & ful",
+            "full-outdoor",
+            "fuel outdoor",
+            "outoor",
+            "outdoor",
+            "outdoor  boiler&full",
+            "boiler outdoor &full",
+            "outdoor boiler &full",
+            "boiler outdoor & ful",
+            "outdoor-boiler",
+            "outdoor - boiler",
+            "outdoor const.",
+            "4 outdoor boilers",
+            "3 outdoor boilers",
+            "full outdoor",
+            "full outdoors",
+            "full oudoors",
+            "outdoor (auto oper)",
+            "outside boiler",
+            "outdoor boiler&full",
+            "outdoor hrsg",
+            "outdoor hrsg",
+            "outdoor-steel encl.",
+            "boiler-outdr & full",
+            "con.& full outdoor",
+            "partial outdoor",
+            "outdoor (auto. oper)",
+            "outdoor (auto.oper)",
+            "outdoor construction",
+            "1 outdoor boiler",
+            "2 outdoor boilers",
+            "outdoor enclosure",
+            "2 outoor boilers",
+            "boiler outdr.& full",
+            "boiler outdr. & full",
+            "ful outdoor",
+            "outdoor-steel enclos",
+            "outdoor (auto oper.)",
+            "con. & full outdoor",
+            "outdore",
+            "boiler & full outdor",
+            "full & outdr boilers",
+            "outodoor (auto oper)",
+            "outdoor steel encl.",
+            "full outoor",
+            "boiler & outdoor ful",
+            "otdr. blr. & f. otdr",
+            "f.otdr & otdr.blr.",
+            "oudoor (auto oper)",
+            "outdoor constructin",
+            "f. otdr. & otdr. blr",
+            "outdoor boiler & fue",
+            "outdoor boiler &fuel",
+        },
+        "semioutdoor": {
+            "semioutdoor" "more than 50% outdoors",
+            "more than 50% outdoo",
+            "more than 50% outdos",
+            "over 50% outdoor",
+            "over 50% outdoors",
+            "semi-outdoor",
+            "semi - outdoor",
+            "semi outdoor",
+            "semi-enclosed",
+            "semi-outdoor boiler",
+            "semi outdoor boiler",
+            "semi- outdoor",
+            "semi - outdoors",
+            "semi -outdoor" "conven & semi-outdr",
+            "conv & semi-outdoor",
+            "conv & semi- outdoor",
+            "convent. semi-outdr",
+            "conv. semi outdoor",
+            "conv(u1)/semiod(u2)",
+            "conv u1/semi-od u2",
+            "conv-one blr-semi-od",
+            "convent semioutdoor",
+            "conv. u1/semi-od u2",
+            "conv - 1 blr semi od",
+            "conv. ui/semi-od u2",
+            "conv-1 blr semi-od",
+            "conven. semi-outdoor",
+            "conv semi-outdoor",
+            "u1-conv./u2-semi-od",
+            "u1-conv./u2-semi -od",
+            "convent. semi-outdoo",
+            "u1-conv. / u2-semi",
+            "conven & semi-outdr",
+            "semi -outdoor",
+            "outdr & conventnl",
+            "conven. full outdoor",
+            "conv. & outdoor blr",
+            "conv. & outdoor blr.",
+            "conv. & outdoor boil",
+            "conv. & outdr boiler",
+            "conv. & out. boiler",
+            "convntl,outdoor blr",
+            "outdoor & conv.",
+            "2 conv., 1 out. boil",
+            "outdoor/conventional",
+            "conv. boiler outdoor",
+            "conv-one boiler-outd",
+            "conventional outdoor",
+            "conventional outdor",
+            "conv. outdoor boiler",
+            "conv.outdoor boiler",
+            "conventional outdr.",
+            "conven,outdoorboiler",
+            "conven full outdoor",
+            "conven,full outdoor",
+            "1 out boil, 2 conv",
+            "conv. & full outdoor",
+            "conv. & outdr. boilr",
+            "conv outdoor boiler",
+            "convention. outdoor",
+            "conv. sem. outdoor",
+            "convntl, outdoor blr",
+            "conv & outdoor boil",
+            "conv & outdoor boil.",
+            "outdoor & conv",
+            "conv. broiler outdor",
+            "1 out boilr, 2 conv",
+            "conv.& outdoor boil.",
+            "conven,outdr.boiler",
+            "conven,outdr boiler",
+            "outdoor & conventil",
+            "1 out boilr 2 conv",
+            "conv & outdr. boilr",
+            "conven, full outdoor",
+            "conven full outdr.",
+            "conven, full outdr.",
+            "conv/outdoor boiler",
+            "convnt'l outdr boilr",
+            "1 out boil 2 conv",
+            "conv full outdoor",
+            "conven, outdr boiler",
+            "conventional/outdoor",
+            "conv&outdoor boiler",
+            "outdoor & convention",
+            "conv & outdoor boilr",
+            "conv & full outdoor",
+            "convntl. outdoor blr",
+            "conv - ob",
+            "1conv'l/2odboilers",
+            "2conv'l/1odboiler",
+            "conv-ob",
+            "conv.-ob",
+            "1 conv/ 2odboilers",
+            "2 conv /1 odboilers",
+            "conv- ob",
+            "conv -ob",
+            "con sem outdoor",
+            "cnvntl, outdr, boilr",
+            "less than 50% outdoo",
+            "less than 50% outdoors",
+            "under 50% outdoor",
+            "under 50% outdoors",
+            "1cnvntnl/2odboilers",
+            "2cnvntnl1/1odboiler",
+            "con & ob",
+            "combination (b)",
+            "indoor & outdoor",
+            "conven. blr. & full",
+            "conv. & otdr. blr.",
+            "combination",
+            "indoor and outdoor",
+            "conven boiler & full",
+            "2conv'l/10dboiler",
+            "4 indor/outdr boiler",
+            "4 indr/outdr boilerr",
+            "4 indr/outdr boiler",
+            "indoor & outdoof",
+        },
+        "conventional": {
+            "conventional",
+            "conventional",
+            "conventional boiler",
+            "conventional - boiler",
+            "conv-b",
+            "conventionall",
+            "convention",
+            "conventional",
+            "coventional",
+            "conven full boiler",
+            "c0nventional",
+            "conventtional",
+            "convential" "underground",
+            "conventional bulb",
+            "conventrional",
+            "*conventional",
+            "convential",
+            "convetional",
+            "conventioanl",
+            "conventioinal",
+            "conventaional",
+            "indoor construction",
+            "convenional",
+            "conventional steam",
+            "conventinal",
+            "convntional",
+            "conventionl",
+            "conventionsl",
+            "conventiional",
+            "convntl steam plants",
+            "indoor const.",
+            "full indoor",
+            "indoor",
+            "indoor automatic",
+            "indoor boiler",
+            "indoor boiler and steam turbine",
+            "(peak load) indoor",
+            "conventionl,indoor",
+            "conventionl, indoor",
+            "conventional, indoor",
+            "conventional;outdoor",
+            "conven./outdoor",
+            "conventional;semi-ou",
+            "comb. cycle indoor",
+            "comb cycle indoor",
+            "3 indoor boiler",
+            "2 indoor boilers",
+            "1 indoor boiler",
+            "2 indoor boiler",
+            "3 indoor boilers",
+            "fully contained",
+            "conv - b",
+            "conventional/boiler",
+            "cnventional",
+            "comb. cycle indooor",
+            "sonventional",
+            "ind enclosures",
+            "conentional",
+            "conventional - boilr",
+            "indoor boiler and st",
+        },
+        "na_category": {
+            "na_category",
+            "na",
+            "",
+            "automatic operation",
+            "comb. turb. installn",
+            "comb. turb. instaln",
+            "com. turb. installn",
+            "n/a",
+            "for detailed info.",
+            "for detailed info",
+            "combined cycle",
+            "not applicable",
+            "gas",
+            "heated individually",
+            "metal enclosure",
+            "pressurized water",
+            "nuclear",
+            "jet engine",
+            "gas turbine",
+            "storage/pipelines",
+            "0",
+            "during 1994",
+            "peaking - automatic",
+            "gas turbine/int. cm",
+            "2 oil/gas turbines",
+            "wind",
+            "package",
+            "mobile",
+            "auto-operated",
+            "steam plants",
+            "other production",
+            "all nuclear plants",
+            "other power gen.",
+            "automatically operad",
+            "automatically operd",
+            "circ fluidized bed",
+            "jet turbine",
+            "gas turbne/int comb",
+            "automatically oper.",
+            "retired 1/1/95",
+            "during 1995",
+            "1996. plant sold",
+            "reactivated 7/1/96",
+            "gas turbine/int comb",
+            "portable",
+            "head individually",
+            "automatic opertion",
+            "peaking-automatic",
+            "cycle",
+            "full order",
+            "circ. fluidized bed",
+            "gas turbine/intcomb",
+            "0.0000",
+            "none",
+            "2 oil / gas",
+            "block & steel",
+            "and 2000",
+            "comb.turb. instaln",
+            "automatic oper.",
+            "pakage",
+            "---",
+            "—",
+            "n/a (ct)",
+            "comb turb instain",
+            "ind encloures",
+            "2 oil /gas turbines",
+            "combustion turbine",
+            "1970",
+            "gas/oil turbines",
+            "combined cycle steam",
+            "pwr",
+            "2 oil/ gas",
+            "2 oil / gas turbines",
+            "gas / oil turbines",
+            "no boiler",
+            "internal combustion",
+            "gasturbine no boiler",
+            "boiler",
+            "tower -10 unit facy",
+            "gas trubine",
+            "4 gas/oil trubines",
+            "2 oil/ 4 gas/oil tur",
+            "5 gas/oil turbines",
+            "tower 16",
+            "2 on 1 gas turbine",
+            "tower 23",
+            "tower -10 unit",
+            "tower - 101 unit",
+            "3 on 1 gas turbine",
+            "tower - 10 units",
+            "tower - 165 units",
+            "wind turbine",
+            "fixed tilt pv",
+            "tracking pv",
+            "o",
+            "wind trubine",
+            "wind generator",
+            "subcritical",
+            "sucritical",
+            "simple cycle",
+            "simple & reciprocat",
+            "solar",
+            "pre-fab power plant",
+            "prefab power plant",
+            "prefab. power plant",
+            "pump storage",
+            "underground",
+            "see page 402",
+            "conv. underground",
+            "conven. underground",
+            "conventional (a)",
+            "non-applicable",
+            "duct burner",
+            "see footnote",
+            "simple and reciprocat",
+        },
+    }
+}
+"""
+A dictionary of construction types (keys) and lists of construction type strings
+associated with each type (values) from FERC Form 1.
+
+There are many strings that weren't categorized, including crosses between conventional
+and outdoor, PV, wind, combined cycle, and internal combustion. The lists are broken out
+into the two types specified in Form 1: conventional and outdoor. These lists are
+inclusive so that variants of conventional (e.g. "conventional full") and outdoor (e.g.
+"outdoor full" and "outdoor hrsg") are included.
+"""
+
+##############################################################################
+# Fully assembled set of FERC 1 transformation parameters
+##############################################################################
+TRANSFORM_PARAMS = {
+    "fuel_ferc1": {
+        "rename_columns_ferc1": {
+            "dbf": {
+                "columns": {
+                    "respondent_id": "utility_id_ferc1",
+                    "plant_name": "plant_name_ferc1",
+                    "fuel": "fuel_type_code_pudl",
+                    "fuel_unit": "fuel_units",
+                    # Original fuel heat content is reported...:
+                    # * coal: almost entirely BTU per POUND
+                    # * gas: ~half MMBTU per cubic foot, ~half MMBTU per Mcf
+                    # * oil: almost entirely BTU per gallon
+                    "fuel_avg_heat": "fuel_btu_per_unit",
+                    "fuel_quantity": "fuel_consumed_units",
+                    "fuel_cost_burned": "fuel_cost_per_unit_burned",
+                    "fuel_cost_delvd": "fuel_cost_per_unit_delivered",
+                    # Note: Original fuel_cost_btu is misleadingly named
+                    "fuel_cost_btu": "fuel_cost_per_mmbtu",
+                    "fuel_generaton": "fuel_btu_per_kwh",
+                    "report_prd": "report_prd",
+                    "row_prvlg": "row_prvlg",
+                    "report_year": "report_year",
+                    "row_number": "row_number",
+                    "fuel_cost_kwh": "fuel_cost_per_kwh",
+                    "spplmnt_num": "spplmnt_num",
+                    "row_seq": "row_seq",
+                }
+            },
+            "xbrl": {
+                "columns": {
+                    "PlantNameAxis": "plant_name_ferc1",
+                    "FuelKindAxis": "fuel_type_code_pudl",
+                    "FuelUnit": "fuel_units",
+                    "FuelBurnedAverageHeatContent": "fuel_btu_per_unit",
+                    "QuantityOfFuelBurned": "fuel_consumed_units",
+                    "AverageCostOfFuelPerUnitBurned": "fuel_cost_per_unit_burned",
+                    "AverageCostOfFuelPerUnitAsDelivered": "fuel_cost_per_unit_delivered",
+                    "AverageCostOfFuelBurnedPerMillionBritishThermalUnit": "fuel_cost_per_mmbtu",
+                    "AverageBritishThermalUnitPerKilowattHourNetGeneration": "fuel_btu_per_kwh",
+                    "AverageCostOfFuelBurnedPerKilowattHourNetGeneration": "fuel_cost_per_kwh",
+                    "ReportYear": "report_year",
+                    "FuelKind": "fuel_kind",
+                    "end_date": "end_date",
+                    "entity_id": "entity_id",
+                    "start_date": "start_date",
+                }
+            },
+        },
+        "categorize_strings": {
+            "fuel_type_code_pudl": FUEL_CATEGORIES,
+            "fuel_units": FUEL_UNIT_CATEGORIES,
+        },
+        "convert_units": {
+            "fuel_btu_per_unit": BTU_TO_MMBTU,
+            "fuel_btu_per_kwh": BTU_PERKWH_TO_MMBTU_PERMWH,
+            "fuel_cost_per_kwh": PERKWH_TO_PERMWH,
+        },
+        "normalize_strings": {
+            "plant_name_ferc1": True,
+            "fuel_type_code_pudl": True,
+            "fuel_units": True,
+        },
+        "correct_units": [
+            {
+                "col": "fuel_mmbtu_per_unit",
+                "query": "fuel_type_code_pudl=='coal'",
+                "valid_range": VALID_COAL_MMBTU_PER_TON,
+                "unit_conversions": [
+                    PERPOUND_TO_PERSHORTTON,
+                    BTU_TO_MMBTU,
+                ],
+            },
+            {
+                "col": "fuel_cost_per_mmbtu",
+                "query": "fuel_type_code_pudl=='coal'",
+                "valid_range": VALID_COAL_USD_PER_MMBTU,
+                "unit_conversions": [
+                    CENTS_TO_DOLLARS,
+                ],
+            },
+            {
+                "col": "fuel_mmbtu_per_unit",
+                "query": "fuel_type_code_pudl=='gas'",
+                "valid_range": VALID_GAS_MMBTU_PER_MCF,
+                "unit_conversions": [
+                    PERCF_TO_PERMCF,
+                    BTU_TO_MMBTU,
+                ],
+            },
+            {
+                "col": "fuel_cost_per_mmbtu",
+                "query": "fuel_type_code_pudl=='gas'",
+                "valid_range": VALID_GAS_USD_PER_MMBTU,
+                "unit_conversions": [
+                    CENTS_TO_DOLLARS,
+                ],
+            },
+            {
+                "col": "fuel_mmbtu_per_unit",
+                "query": "fuel_type_code_pudl=='oil'",
+                "valid_range": VALID_OIL_MMBTU_PER_BBL,
+                "unit_conversions": [
+                    PERGALLON_TO_PERBARREL,
+                    BTU_TO_MMBTU,  # Why was this omitted in the old corrections?
+                ],
+            },
+            {
+                "col": "fuel_cost_per_mmbtu",
+                "query": "fuel_type_code_pudl=='oil'",
+                "valid_range": VALID_OIL_USD_PER_MMBTU,
+                "unit_conversions": [
+                    CENTS_TO_DOLLARS,
+                ],
+            },
+        ],
+    },
+    "plants_steam_ferc1": {
+        "normalize_strings": {
+            "plant_name_ferc1": True,
+            "construction_type": True,
+            "plant_type": True,
+        },
+        "nullify_outliers": {
+            "construction_year": VALID_PLANT_YEARS,
+            "installation_year": VALID_PLANT_YEARS,
+        },
+        "categorize_strings": {
+            "construction_type": CONSTRUCTION_TYPE_CATEGORIES,
+            "plant_type": PLANT_TYPE_CATEGORIES,
+        },
+        "convert_units": {
+            "capex_per_kw": PERKW_TO_PERMW,
+            "opex_per_kwh": PERKWH_TO_PERMWH,
+            "net_generation_kwh": KWH_TO_MWH,
+        },
+        "rename_columns_ferc1": {
+            "dbf": {
+                "columns": {
+                    "cost_structure": "capex_structures",
+                    "expns_misc_power": "opex_misc_power",
+                    "plant_name": "plant_name_ferc1",
+                    "plnt_capability": "plant_capability_mw",
+                    "expns_plants": "opex_plants",
+                    "expns_misc_steam": "opex_misc_steam",
+                    "cost_per_kw": "capex_per_kw",
+                    "when_not_limited": "not_water_limited_capacity_mw",
+                    "asset_retire_cost": "asset_retirement_cost",
+                    "expns_steam_othr": "opex_steam_other",
+                    "expns_transfer": "opex_transfer",
+                    "expns_engnr": "opex_engineering",
+                    "avg_num_of_emp": "avg_num_employees",
+                    "cost_of_plant_to": "capex_total",
+                    "expns_rents": "opex_rents",
+                    "tot_prdctn_expns": "opex_production_total",
+                    "plant_kind": "plant_type",
+                    "respondent_id": "utility_id_ferc1",
+                    "expns_operations": "opex_operations",
+                    "cost_equipment": "capex_equipment",
+                    "type_const": "construction_type",
+                    "plant_hours": "plant_hours_connected_while_generating",
+                    "expns_coolants": "opex_coolants",
+                    "expns_fuel": "opex_fuel",
+                    "when_limited": "water_limited_capacity_mw",
+                    "expns_kwh": "opex_per_kwh",
+                    "expns_allowances": "opex_allowances",
+                    "expns_steam": "opex_steam",
+                    "yr_const": "construction_year",
+                    "yr_installed": "installation_year",
+                    "expns_boiler": "opex_boiler",
+                    "peak_demand": "peak_demand_mw",
+                    "cost_land": "capex_land",
+                    "tot_capacity": "capacity_mw",
+                    "net_generation": "net_generation_kwh",
+                    "expns_electric": "opex_electric",
+                    "expns_structures": "opex_structures",
+                    "report_year": "report_year",
+                    "report_prd": "report_prd",
+                    "row_prvlg": "row_prvlg",
+                    "row_number": "row_number",
+                    "spplmnt_num": "spplmnt_num",
+                    "row_seq": "row_seq",
+                }
+            },
+            "xbrl": {
+                "columns": {
+                    "CostOfStructuresAndImprovementsSteamProduction": "capex_structures",
+                    "MiscellaneousSteamPowerExpenses": "opex_misc_power",
+                    "PlantNameAxis": "plant_name_ferc1",
+                    "NetContinuousPlantCapability": "plant_capability_mw",
+                    "MaintenanceOfElectricPlantSteamPowerGeneration": "opex_plants",
+                    "MaintenanceOfMiscellaneousSteamPlant": "opex_misc_steam",
+                    "CostPerKilowattOfInstalledCapacity": "capex_per_kw",
+                    "NetContinuousPlantCapabilityNotLimitedByCondenserWater": "not_water_limited_capacity_mw",
+                    "AssetRetirementCostsSteamProduction": "asset_retirement_cost",
+                    "SteamFromOtherSources": "opex_steam_other",
+                    "SteamTransferredCredit": "opex_transfer",
+                    "MaintenanceSupervisionAndEngineeringSteamPowerGeneration": "opex_engineering",
+                    "PlantAverageNumberOfEmployees": "avg_num_employees",
+                    "CostOfPlant": "capex_total",
+                    "RentsSteamPowerGeneration": "opex_rents",
+                    "PowerProductionExpensesSteamPower": "opex_production_total",
+                    "PlantKind": "plant_type",
+                    "OperationSupervisionAndEngineeringExpense": "opex_operations",
+                    "CostOfEquipmentSteamProduction": "capex_equipment",
+                    "PlantConstructionType": "construction_type",
+                    "PlantHoursConnectedToLoad": "plant_hours_connected_while_generating",
+                    "CoolantsAndWater": "opex_coolants",
+                    "FuelSteamPowerGeneration": "opex_fuel",
+                    "NetContinuousPlantCapabilityLimitedByCondenserWater": "water_limited_capacity_mw",
+                    "ExpensesPerNetKilowattHour": "opex_per_kwh",
+                    "Allowances": "opex_allowances",
+                    "SteamExpensesSteamPowerGeneration": "opex_steam",
+                    "YearPlantOriginallyConstructed": "construction_year",
+                    "YearLastUnitOfPlantInstalled": "installation_year",
+                    "MaintenanceOfBoilerPlantSteamPowerGeneration": "opex_boiler",
+                    "NetPeakDemandOnPlant": "peak_demand_mw",
+                    "CostOfLandAndLandRightsSteamProduction": "capex_land",
+                    "InstalledCapacityOfPlant": "capacity_mw",
+                    "NetGenerationExcludingPlantUse": "net_generation_kwh",
+                    "ElectricExpensesSteamPowerGeneration": "opex_electric",
+                    "MaintenanceOfStructuresSteamPowerGeneration": "opex_structures",
+                    "ReportYear": "report_year",
+                }
+            },
+        },
+    },
+}
